fix: webhook_count BIGINT type mismatch — i32 → i64
Migration 011 changed webhook_count from INT to BIGINT but the
Customer struct still used i32. sqlx type deserialization fails
when reading BIGINT into i32, causing DATABASE_ERROR on all
customer queries (login, register, etc).

Changes:
- Customer.webhook_count: i32 → i64
- CustomerResponse.webhook_count: i32 → i64
- ProfileResponse.webhook_count: i32 → i64
- AdminUserDetail.webhook_count: i32 → i64
- Test values updated (i32::MAX → i64::MAX)

+++ api/src/routes/customer_portal.rs
@@ -41,7 +41,7 @@ struct ProfileResponse {
     email: String,
     plan: String,
     webhook_limit: i32,
-    webhook_count: i32,
+    webhook_count: i64,
     created_at: String,
 }
